# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "internet-date API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/profiles/{userId}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get profile",
        "security": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "profiles"
        ],
        "summary": "Update own profile",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilePatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/{userId}/rate": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Rate profile",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^\\d+$"
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "PUBLIC",
          "PRIVATE"
        ]
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "IMAGE",
          "VIDEO"
        ]
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "UNSPECIFIED",
          "MALE",
          "FEMALE",
          "NONBINARY",
          "OTHER"
        ]
      },
      "DatingIntent": {
        "type": "string",
        "enum": [
          "UNSPECIFIED",
          "FRIENDS",
          "CASUAL",
          "LONG_TERM",
          "MARRIAGE"
        ]
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "url": {
            "type": "string"
          },
          "thumbUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "number",
              "null"
            ]
          },
          "height": {
            "type": [
              "number",
              "null"
            ]
          },
          "durationSec": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "type",
          "url"
        ]
      },
...
      "PostMedia": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number"
          },
          "media": {
            "$ref": "#/components/schemas/Media"
          }
        },
        "required": [
          "order",
          "media"
        ]
      },
...
      "ProfileTop5Item": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "order",
          "text"
        ]
      },
      "ProfileTop5List": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileTop5Item"
            }
          }
        },
        "required": [
          "id",
          "title",
          "updatedAt",
          "items"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "birthdate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "locationText": {
            "type": [
              "string",
              "null"
            ]
          },
          "gender": {
            "type": [
              "string",
              "null"
            ]
          },
          "intent": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "heroUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "isVisible": {
            "type": "boolean"
          },
          "top5Lists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileTop5List"
            }
          }
        },
        "required": [
          "userId",
          "isVisible",
          "top5Lists"
        ]
      },
      "ProfilePost": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMedia"
            }
          }
        },
        "required": [
          "id",
          "visibility",
          "createdAt",
          "media"
        ]
      },
      "ProfileRatingAvg": {
        "type": "object",
        "properties": {
          "attractive": {
            "type": [
              "number",
              "null"
            ]
          },
          "smart": {
            "type": [
              "number",
              "null"
            ]
          },
          "funny": {
            "type": [
              "number",
              "null"
            ]
          },
          "interesting": {
            "type": [
              "number",
              "null"
            ]
          }
        }
      },
      "ProfileRatingMine": {
        "type": "object",
        "properties": {
          "attractive": {
            "type": "number"
          },
          "smart": {
            "type": "number"
          },
          "funny": {
            "type": "number"
          },
          "interesting": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "attractive",
          "smart",
          "funny",
          "interesting",
          "createdAt"
        ]
      },
      "ProfileRatings": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          },
          "avg": {
            "$ref": "#/components/schemas/ProfileRatingAvg"
          },
          "mine": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileRatingMine"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "count",
          "avg",
          "mine"
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Profile"
              },
              {
                "type": "null"
              }
            ]
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfilePost"
            }
          },
          "ratings": {
            "$ref": "#/components/schemas/ProfileRatings"
          }
        },
        "required": [
          "profile",
          "posts",
          "ratings"
        ]
      },
      "ProfilePatchBody": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "birthdate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "locationText": {
            "type": [
              "string",
              "null"
            ]
          },
          "lat": {
            "type": [
              "number",
              "null"
            ]
          },
          "lng": {
            "type": [
              "number",
              "null"
            ]
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "intent": {
            "$ref": "#/components/schemas/DatingIntent"
          },
          "isVisible": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "avatarMediaId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          },
          "heroMediaId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Id"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ProfilePatchResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Id"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "updatedAt"
        ]
      },
      "RateBody": {
        "type": "object",
        "properties": {
          "attractive": {
            "type": "number"
          },
          "smart": {
            "type": "number"
          },
          "funny": {
            "type": "number"
          },
          "interesting": {
            "type": "number"
          }
        },
        "required": [
          "attractive",
          "smart",
          "funny",
          "interesting"
        ]
      },
...
    }
  }
}